Revert "feat: added line style customization per property"
This reverts commit 5698be9ebfaf2ad47dadbc5c03b21022654486ab.
