+++ packages/dashboard/src/customization/propertiesSections/propertiesAndAlarmsSettings/styledPropertyComponent.tsx
@@ -1,5 +1,4 @@
 import React, { useEffect, useRef, useState } from 'react';
-import type { FC } from 'react';
 import {
   Button,
   SpaceBetween,
@@ -10,71 +9,20 @@ import {
   FormField,
   Input,
 } from '@cloudscape-design/components';
-import { spaceStaticXxs } from '@cloudscape-design/design-tokens';
-import { ClickDetail, NonCancelableEventHandler } from '@cloudscape-design/components/internal/events';
-
 import ColorPicker from '../shared/colorPicker';
-import { LineStyles, StyledAssetPropertyQuery, YAxisOptions } from '~/customization/widgets/types';
+import type { FC } from 'react';
+
+import './propertyComponent.css';
+import { StyledAssetPropertyQuery, YAxisOptions } from '~/customization/widgets/types';
 import { getPropertyDisplay } from './getPropertyDisplay';
 import type { AssetSummary } from '~/hooks/useAssetDescriptionQueries';
 import {
   StatusEyeHidden,
   StatusEyeVisible,
 } from '~/customization/propertiesSections/propertiesAndAlarmsSettings/icons';
+import { ClickDetail, NonCancelableEventHandler } from '@cloudscape-design/components/internal/events';
 import Tooltip from '~/components/tooltip/tooltip';
-import { LineStyleDropdown } from '../components/lineStyleDropdown';
-import { LineThicknessDropdown } from '../components/lineThicknessDropdown';
-
-import './propertyComponent.css';
-
-const LineStylePropertyConfig = ({
-  resetStyles,
-  property,
-  onUpdate,
-}: {
-  resetStyles: (reset: object) => void;
-  property: StyledAssetPropertyQuery;
-  onUpdate: (newStyles: object) => void;
-}) => {
-  const [expanded, setExpanded] = useState(false);
-  const [useGlobalStyle, setUseGlobalStyle] = useState<boolean>(!property.line?.style); //make useGlobalStyle true if no line style
-  const [lineStyle, setLineStyle] = useState<LineStyles['style']>(property.line?.style ?? 'solid');
-
-  const onToggleUseGlobalStyles = (isChecked: boolean) => {
-    setUseGlobalStyle(isChecked);
-    setLineStyle('solid');
-    if (isChecked) {
-      resetStyles({ line: { style: undefined } });
-    } else {
-      onUpdate({ line: { style: lineStyle } });
-    }
-  };
-
-  const updateLineStyle = (style: LineStyles['style']) => {
-    setLineStyle(style);
-    onUpdate({ line: { style } });
-  };
-
-  return (
-    <ExpandableSection
-      expanded={expanded}
-      onChange={({ detail }) => setExpanded(detail.expanded)}
-      headerText='Line style'
-    >
-      <SpaceBetween size='m'>
-        <Checkbox onChange={({ detail }) => onToggleUseGlobalStyles(detail.checked)} checked={useGlobalStyle}>
-          Use default style
-        </Checkbox>
-        <LineStyleDropdown
-          disabled={useGlobalStyle}
-          lineStyle={lineStyle}
-          updatelineStyle={(style) => updateLineStyle(style as LineStyles['style'])}
-        />
-        <LineThicknessDropdown disabled={useGlobalStyle} /> {/* TODO */}
-      </SpaceBetween>
-    </ExpandableSection>
-  );
-};
+import { spaceStaticXxs } from '@cloudscape-design/design-tokens';
 
 const numberOrUndefined = (number: string) => {
   const parsed = Number(number);
@@ -225,7 +173,6 @@ export const StyledPropertyComponent: FC<StyledPropertyComponentProps> = ({
             <SpaceBetween size='xs' direction='horizontal'>
               <ExpandableSection headerText={YAxisHeader}>
                 <div style={{ padding: '0 24px', backgroundColor: '#fbfbfb' }}>
-                  <LineStylePropertyConfig resetStyles={resetStyles} onUpdate={updateStyle} property={property} />
                   <YAxisPropertyConfig resetStyles={resetStyles} onUpdate={updateStyle} property={property} />
                 </div>
               </ExpandableSection>
